cleaner APIs; better proc spawning and termination
Made it possible to pass `ctx` as-is to various APIs that take an
"options" argument with a `.signal`, or spread it into another options
object.

Removed the need for passing `ctx` as the first argument to various
functions. Passing it to `spawn` and `fork` is now optional, though
recommended.

Moved `par` and `ser` to `Ctx` as methods, for consistency with `run`
and to reduce mistypings.

Process-spawning functions now have the same signature as the Node
functions they wrap. `spawn` and `fork` can take additional options.

Register processes for additional cleanup via `process.once('exit')`.
On Unix, this allows us to terminate subprocess groups before Node
terminates for any reason other than VM crash, which appears to solve
the problem of dangling sub-sub-processes. On Windows, this has no
effect.

Updated and improved docs.